requirements.txt
fastapi
uvicorn
openai
python-dotenv
azure-ai-inference